Updated Opencast backend with latest annotation webapp version

+++ opencast-backend/entwine-annotations-tool/src/main/resources/ui/tests/js/require/config-loading.js
@@ -8,7 +8,7 @@ require.config({
         "bootstrap"                      : "libs/bootstrap/bootstrap.min",
         "carousel"                       : "libs/bootstrap/carousel2.2",
         "domReady"                       : "libs/require/config/domReady",
-        "handlebars"                     : "libs/handlebars-v1.1.2",
+        "handlebars"                     : "libs/handlebars-v1.3.0",
         "handlebarsHelpers"              : "handlebarsHelpers",
         "jquery.colorPicker"             : "libs/jquery.colorPicker.min",
         "jquery.FileReader"              : "libs/jquery.FileReader",
@@ -24,8 +24,9 @@ require.config({
         "templates"                      : "../templates",
         "text"                           : "libs/require/config/text",
         "tooltip"                        : "libs/bootstrap/tooltip",
-        "timeline"                       : "libs/timeline-min",
-        "underscore"                     : "libs/underscore-min-1.4.3"
+        "timeline"                       : "libs/timeline",
+        "underscore"                     : "libs/underscore-min-1.4.3",
+        "raf"                            : "libs/rAF"
     },
     waitSeconds: 10,
 
@@ -77,45 +78,27 @@ require.config({
 });
 
 // Bootstrap function for main app
-require(["domReady", "annotations-tool-configuration", "annotations-tool"],
+require(["domReady", "annotations-tool-configuration", "annotations-tool", "raf"],
 
 function (domReady, config, app) {
     domReady(function () {
 
-        // var lastUpdate = 0,
-        //     i = 0,
-        //     nbFrame = 20,
-        //     lastValuesSum = 0,
-        //     targetFPS = $("#fps #value"),
-        //     calculateFPS = function () {
-        //         var currentTime = (new Date()).getTime();
+        var getURLParameter = function (name) {
+                                return decodeURIComponent((new RegExp("[?|&]" + name + "=" + "([^&;]+?)(&|#|;|$)").exec(location.search) ||
+                                    [,""])[1].replace(/\+/g, "%20")) || null;
+                            };
 
-        //         lastValuesSum += (currentTime - lastUpdate);
-
-        //         if (++i === nbFrame) {
-        //             targetFPS.html(1000 / (lastValuesSum / nbFrame));
-        //             i = 0;
-        //             lastValuesSum = 0;
-        //         }
-
-        //         lastUpdate = currentTime;
-        //     };
-
-        localStorage.clear();
+        if (getURLParameter("reset") !== null || typeof localStorage.Users === "undefined") {
+            localStorage.clear();
+        } else {
+            config.tracksToImport = undefined;
+        }
 
         localStorage.Users = "c1";
         localStorage["Users-c1"] = "{\"user_extid\": \"default\",\"nickname\": \"test\",\"role\": \"user\",\"access\": 1,\"id\": \"c1\",\"created_at\": null,\
                                     \"updated_at\": null,\"deleted_at\": null,\"email\": \"test@test.ch\"}";
 
         app.start(config);
-
-        // if (_.isUndefined(annotationsTool.video)) {
-        //     annotationsTool.bind(annotationsTool.EVENTS.READY, function () {
-        //         annotationsTool.bind(annotationsTool.EVENTS.TIMEUPDATE, calculateFPS);
-        //     });
-        // } else {
-        //     annotationsTool.bind(annotationsTool.EVENTS.TIMEUPDATE, calculateFPS);
-        // }
 
     });
 });
